+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/config/Cdoc2RpClientConfiguration.java
@@ -2,6 +2,8 @@ package ee.cyber.cdoc2.config;
 
 import java.util.Properties;
 
+import ee.cyber.cdoc2.client.model.MidDisplayTextFormat;
+import ee.cyber.cdoc2.client.model.MidLanguage;
 import ee.cyber.cdoc2.exceptions.ConfigurationLoadingException;
 
 public interface Cdoc2RpClientConfiguration {
@@ -12,10 +14,29 @@ public interface Cdoc2RpClientConfiguration {
     }
 
     String getHostUrl();
+
     String getCertificateLevel();
+
     String getTrustStore();
+
     String getTrustStorePassword();
-    String getAuthenticationTrustStore();
-    String getAuthenticationTrustStorePassword();
+
+    String getSidSigningCertificateTrustStore();
+
+    String getSidSigningCertificateTrustStorePassword();
+
+    String getMidSigningCertificateTrustStore();
+
+    String getMidSigningCertificateTrustStorePassword();
+
+    /**
+     * Default display text, can be overwritten with InteractionParams
+     */
+    String getDefaultDisplayText();
+
+    MidDisplayTextFormat getDefaultDisplayTextFormat();
+
+    MidLanguage getDefaultDisplayTextLanguage();
+
     boolean getClientServerDebug();
 }