+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/crypto/jwt/MIDAuthJWSSigner.java
@@ -1,43 +1,52 @@
 package ee.cyber.cdoc2.crypto.jwt;
 
+import ee.sk.mid.MidAuthenticationHashToSign;
+import ee.sk.mid.MidHashToSign;
+import ee.sk.mid.MidHashType;
+import ee.sk.mid.exception.MidInvalidNationalIdentityNumberException;
+import ee.sk.mid.exception.MidInvalidPhoneNumberException;
+import jakarta.annotation.Nullable;
+
+import java.security.cert.X509Certificate;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jose.jca.JCAContext;
 import com.nimbusds.jose.util.Base64URL;
+import com.nimbusds.jose.util.X509CertUtils;
 
 import ee.cyber.cdoc2.auth.EtsiIdentifier;
-import ee.cyber.cdoc2.client.mobileid.MobileIdClient;
 import ee.cyber.cdoc2.client.mobileid.MobileIdUserData;
-import ee.cyber.cdoc2.exceptions.CdocMobileIdClientException;
-import ee.sk.mid.MidAuthentication;
-import ee.sk.mid.MidAuthenticationHashToSign;
-import ee.sk.mid.MidHashToSign;
-import ee.sk.mid.MidHashType;
-
-import ee.sk.mid.exception.MidInvalidNationalIdentityNumberException;
-import ee.sk.mid.exception.MidInvalidPhoneNumberException;
-import jakarta.annotation.Nullable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.security.cert.X509Certificate;
-import java.util.Objects;
-import java.util.Set;
+import ee.cyber.cdoc2.client.model.MidSessionStatusResponse;
+import ee.cyber.cdoc2.client.rpserver.Cdoc2RpClient;
+import ee.cyber.cdoc2.exceptions.CdocRpClientException;
 
 /**
  * JWSSigner that implements signing using Mobile-ID authentication key/certificate. Supports only ES256 algorithm.
  * At REST API level signer is identified by "phone number" and "identity code" which is not SematicsIdentifier.
+ *
  * @ see <a href="https://github.com/SK-EID/MID">Mobile ID (MID) REST API</a>
  */
 public class MIDAuthJWSSigner implements IdentityJWSSigner {
+    private static final TimeUnit SESSION_POLL_SLEEP_TIMEUNIT = TimeUnit.SECONDS;
+    private static final long SESSION_POLL_SLEEP_QUANTITY = 1L;
 
     private static final Logger log = LoggerFactory.getLogger(MIDAuthJWSSigner.class);
     private final JCAContext jcaContext = new JCAContext();
 
-    private final MobileIdClient midClient;
+    private final Cdoc2RpClient rpClient;
     private final EtsiIdentifier signerEtsiIdentifier;
     private final MobileIdUserData mobileIdUserData;
+    private final SessionToken sessionToken;
 
     private final @Nullable InteractionParams interactionParams;
 
@@ -46,23 +55,30 @@ public class MIDAuthJWSSigner implements IdentityJWSSigner {
     /**
      * Initialize JWSSigner with MobileIdClient and signer identified by identity code and phone number
      * and pre-initialized MobileIdClient
-     * @param midClient MobileIdClient to perform actual authentication sequence
-     * @param signer signer identifier as etsi semantics identifier
-     * @param phoneNumber signer phone number in international format e.g. "+3725551234"
-     * @param interactionParams  Optional parameters to drive user interaction. {@code null} if not used
-     * @throws MidInvalidPhoneNumberException if phone number validation has failed
+     *
+     * @param rpClient          RP client to perform actual authentication sequence
+     * @param signer            signer identifier as etsi semantics identifier
+     * @param phoneNumber       signer phone number in international format e.g. "+3725551234"
+     * @param interactionParams Optional parameters to drive user interaction. {@code null} if not used
+     * @throws MidInvalidPhoneNumberException            if phone number validation has failed
      * @throws MidInvalidNationalIdentityNumberException if ID code validation has failed
      */
-    public MIDAuthJWSSigner(EtsiIdentifier signer, String phoneNumber, MobileIdClient midClient,
-                            @Nullable InteractionParams interactionParams) {
-        Objects.requireNonNull(midClient);
+    public MIDAuthJWSSigner(
+        EtsiIdentifier signer,
+        String phoneNumber,
+        Cdoc2RpClient rpClient,
+        @Nullable InteractionParams interactionParams,
+        SessionToken sessionToken
+    ) {
+        Objects.requireNonNull(rpClient);
         Objects.requireNonNull(signer);
         Objects.requireNonNull(phoneNumber);
 
-        this.midClient = midClient;
+        this.rpClient = rpClient;
         this.signerEtsiIdentifier = signer;
         this.mobileIdUserData = new MobileIdUserData(phoneNumber, signer.getIdentifier());
         this.interactionParams = interactionParams;
+        this.sessionToken = sessionToken;
     }
 
     @Override
@@ -76,7 +92,8 @@ public class MIDAuthJWSSigner implements IdentityJWSSigner {
             throw new JOSEException("JWSAlgorithm " + jwsAlg + " not supported");
         }
 
-        MidAuthenticationHashToSign hash = calcHash(signingInput, toMIDHashType(jwsAlg));
+        MidHashType midHashType = toMIDHashType(jwsAlg);
+        MidAuthenticationHashToSign hash = calcHash(signingInput, midHashType);
 
         if (interactionParams != null) {
             AuthEvent authEvent = new AuthEvent(this, hash.calculateVerificationCode(),
@@ -87,13 +104,61 @@ public class MIDAuthJWSSigner implements IdentityJWSSigner {
         }
 
         try {
-            MidAuthentication result = midClient.startAuthentication(
-                mobileIdUserData, hash, interactionParams);
-            this.signerCertificate = result.getCertificate();
-            return Base64URL.encode(result.getSignatureValue());
-        } catch (CdocMobileIdClientException ex) {
+            String disclosedSessionToken = sessionToken.getSessionToken(rpClient.getBaseUrl());
+
+            UUID sessionId = rpClient.midAuthenticate(
+                disclosedSessionToken,
+                sessionToken.getSigningCertificate(),
+                mobileIdUserData.identityCode(),
+                mobileIdUserData.phoneNumber(),
+                hash.getHash(),
+                midHashType.toString(),
+                interactionParams
+            );
+
+            MidSessionStatusResponse response = pollForFinalSessionStatus(
+                disclosedSessionToken,
+                sessionToken.getSigningCertificate(),
+                sessionId
+            );
+
+            String result = Optional.ofNullable(response.getResult())
+                .map(MidSessionStatusResponse.ResultEnum::getValue)
+                .orElse(null);
+
+            if (!"OK".equals(result)) {
+                String message = "SID session endResult: " + result;
+                log.error(message);
+                throw new CdocRpClientException(message);
+            }
+
+            this.signerCertificate = X509CertUtils.parse(response.getCert());
+
+            return Base64URL.encode(response.getSignature().getValue());
+        } catch (CdocRpClientException ex) {
             throw new JOSEException(ex);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
         }
+    }
+
+    private MidSessionStatusResponse pollForFinalSessionStatus(
+        String xCdoc2SessionToken,
+        String xCdoc2SessionX5c,
+        UUID sessionId
+    ) throws InterruptedException, CdocRpClientException {
+        MidSessionStatusResponse sessionStatus = null;
+        while (sessionStatus == null || "RUNNING".equalsIgnoreCase(sessionStatus.getState().getValue())) {
+            sessionStatus = rpClient.midSession(xCdoc2SessionToken, xCdoc2SessionX5c, sessionId);
+            if (sessionStatus != null && "COMPLETE".equalsIgnoreCase(sessionStatus.getState().getValue())) {
+                break;
+            }
+            log.debug("Sleeping for {} {}", SESSION_POLL_SLEEP_QUANTITY,
+                SESSION_POLL_SLEEP_TIMEUNIT);
+            SESSION_POLL_SLEEP_TIMEUNIT.sleep(SESSION_POLL_SLEEP_QUANTITY);
+        }
+        log.debug("Got final session status response");
+        return sessionStatus;
     }
 
     @Override
@@ -103,6 +168,7 @@ public class MIDAuthJWSSigner implements IdentityJWSSigner {
 
     /**
      * After {@link #sign(JWSHeader, byte[])} has succeeded, signer public certificate can be queried
+     *
      * @return signer certificate if {@code sign()} has succeeded, otherwise will be {@code null}
      */
     public @Nullable X509Certificate getSignerCertificate() {
@@ -111,6 +177,7 @@ public class MIDAuthJWSSigner implements IdentityJWSSigner {
 
     /**
      * Not used for MID signatures
+     *
      * @return null
      */
     @Nullable